Refactor config use
