+++ .gitignore
@@ -7,6 +7,7 @@ dist
 build
 .nyc_output
 
+.DS_Store
 temp
 sigmate-v3-ti.json
 src/creds.js
