+++ src/EventStore/Factory/EventStoreFactory.spec.ts
