+++ src/EventStore/Factory/EventStoreFactory.ts
