+++ src/TransmuteFramework.spec.ts
@@ -33,5 +33,12 @@ describe('TransmuteFramework', () => {
             TransmuteFramework.init()
             assert(TransmuteFramework.TransmuteIpfs.config.host === 'localhost')
         })
+
+        it('should initialize EventStore with config', async () => {
+            TransmuteFramework.init()
+            assert(TransmuteFramework.EventStore.framework.TransmuteIpfs.config.host === 'localhost')
+        })
     })
+
+
 })
