+++ src/TransmuteFramework.ts
@@ -1,6 +1,10 @@
 
 import { EventStore } from './EventStore/EventStore'
 
+import { EventTypes } from './EventStore/EventTypes/EventTypes'
+
+import { Persistence } from './EventStore/Persistence/Persistence'
+
 import { TransmuteIpfs, ITransmuteIpfs } from './TransmuteIpfs/TransmuteIpfs'
 
 const Web3 = require('web3')
@@ -31,18 +35,22 @@ export interface ITransmuteFramework {
   EventStore: any,
   init: (confObj?: any) => any,
   config: any,
-  web3: any
+  web3: any,
+  TransmuteIpfs: any,
+  Persistence: any
 }
 
 export class TransmuteFramework implements ITransmuteFramework {
 
   EventStoreFactoryContract = null
   EventStoreContract = null
-  EventStore = EventStore
+  EventStore: any = null
   config = config
   web3 = null
 
   TransmuteIpfs: ITransmuteIpfs = TransmuteIpfs
+  EventTypes = EventTypes
+  Persistence = Persistence
 
   public init = (confObj = config) => {
     this.config = confObj
@@ -69,7 +77,9 @@ export class TransmuteFramework implements ITransmuteFramework {
         protocol: 'http'
       }
     }
+    // This is initialized like so because it can be useful outside framework...
     this.TransmuteIpfs = TransmuteIpfs.init(ipfsConfig)
+    this.EventStore = new EventStore(this)
     return this
   }
